SWBIT-4217 #comment Creates LoginForm component
